iCalendar Reader: Re-order logic to avoid procession and incorrect incrementing.
Both of these problems came up in the default processing of events, so fixing them involved a more extensive rewrite to both simplify the calculations of DTEND for all events and make a faster jump over the interval from an event's original date.

Procession occurs when a timezone-adjusted time is adjusted again, moving the clock around by the timezone offset for an event.

To prevent procession, all DTEND dates are calculated from a time-adjusted DTSTART plus a duration that is saved while setting up the event for looping.

Incrementing for intervals needs to take place from the original start date of an event without processing; otherwise, we have events falling out of place.

Day and week intervals are now incremented using the constants DAY_IN_SECONDS and WEEK_IN_SECONDS. The difference between now and the original start date is divided by the appropriate constant, and the floor of that operation is used to calculate a more recent point in the past that falls on the interval. This avoids looping up from the original event for daily and weekly events.

`strtotime()` is used to add month intervals to the original start date, tracking the number of intervals until a date in the future is reached. The time is then moved forward to find the next valid occurrence of the event.

Merges r122743-wpcom.