iCalendar Reader: Fix multi-day event display.
An all-day event is flagged as starting at the beginning of one day and lasting until the beginning of the last date. In human terms, we understand this as meaning "until", rather than "including". On multi-day events, the last date is incorrectly displayed in our date range, rather than presenting the event as taking place on the previous day.

Subtracting one minute from the time moves the date back to the last inclusive date for a given event, which is comparable to the event display in Google's Calendar and Apple's Calendar app.

Merges r123662-wpcom.