WPCOM: bandcamp shortcode: add support for 'tracks' and 'esig' attributes.
From Robbie Scott <robbie@bandcamp.com> -

We are going to support "exclusive embedded players" soon, meaning an artist can specify a domain (eg, "stereogum.com") that a music can player can be embedded in, exclusively, and the tracks will stream there and only there.

In order to support this, I've added two parameters: "tracks", which is a comma-separated list of integers, and "esig", which is a (hex) md5 digest.

Merges r126424-wpcom.